+++ inc/customizer/custom-color-theme.php
@@ -233,13 +233,12 @@ function backbone_get_base_theme_choices() {
 function backbone_customize_preview_custom_colors() {
     // 管理画面キャッシュバスティング設定を取得
     $cache_busting_admin = get_theme_mod('enable_cache_busting_admin', false);
-    $version_admin = $cache_busting_admin ? current_time('YmdHis') : wp_get_theme()->get('Version');
 
     wp_enqueue_script(
         'seo-optimus-customize-preview-custom-colors',
         get_template_directory_uri() . '/js/customize-preview-custom-colors.js',
         array('customize-preview', 'jquery'),
-        $version_admin,
+        backbone_get_file_version('/js/customize-preview-custom-colors.js', $cache_busting_admin),
         true
     );
 
@@ -255,13 +254,12 @@ add_action('customize_preview_init', 'backbone_customize_preview_custom_colors')
 function backbone_customize_controls_custom_colors() {
     // 管理画面キャッシュバスティング設定を取得
     $cache_busting_admin = get_theme_mod('enable_cache_busting_admin', false);
-    $version_admin = $cache_busting_admin ? current_time('YmdHis') : wp_get_theme()->get('Version');
 
     wp_enqueue_script(
         'seo-optimus-customize-controls-custom-colors',
         get_template_directory_uri() . '/js/customize-controls-custom-colors.js',
         array('customize-controls', 'jquery'),
-        $version_admin,
+        backbone_get_file_version('/js/customize-controls-custom-colors.js', $cache_busting_admin),
         true
     );